Move device parsing to NimbleDeviceListener base class
