+++ esphome_nimble/components/nimble_tracker/irk_utils.cpp
@@ -3,7 +3,7 @@
 
 namespace esphome
 {
-    namespace nimble_rssi
+    namespace nimble_tracker
     {
         int bt_encrypt_be(const uint8_t *key, const uint8_t *plaintext, uint8_t *enc_data)
         {
@@ -56,6 +56,6 @@ namespace esphome
             return true;
         }
 
-    } // namespace nimble_rssi
+    } // namespace nimble_tracker
 
 } // namespace esphome
