+++ esphome_nimble/components/nimble_tracker/irk_utils.h
@@ -6,7 +6,7 @@
 
 namespace esphome
 {
-    namespace nimble_rssi
+    namespace nimble_tracker
     {
         int bt_encrypt_be(const uint8_t *key, const uint8_t *plaintext, uint8_t *enc_data);
 
@@ -19,6 +19,6 @@ namespace esphome
 
         bool ble_ll_resolv_rpa(const uint8_t *rpa, const uint8_t *irk);
 
-    } // namespace nimble_rssi
+    } // namespace nimble_tracker
 
 } // namespace esphome
