+++ esphome_nimble/components/nimble_tracker/string_utils.h
@@ -6,13 +6,13 @@
 
 namespace esphome
 {
-    namespace nimble_rssi
+    namespace nimble_tracker
     {
 
         std::string hexStr(const uint8_t *data, int len);
         uint8_t hextob(char ch);
         bool hextostr(const std::string &hexStr, uint8_t *output, size_t len);
 
-    } // namespace nimble_rssi
+    } // namespace nimble_tracker
 
 } // namespace esphome
